Fix avatar on custom domains
User.avatar_url returns a full URL including protocol and hostname. The
included hostname is always the primary address. It does not respect the
custom domain of the current event, because the property does not have
access to that information. This adapts the CSP to allow the request to
the pretalx main domain, even if the client is on the custom domain.